state machine rework complete
New state machine rework is complete with a new set of quality of life changes!

Since enemy prefabs are not updated to reflect the changes (nor do I plan on updating them until I finish the enemy redesign), enemies are unstable and may or may not be working on this branch as of this commit.