added tile grid update on destroy
Tile grid now properly updates when a tile is destroyed. Can test by enabling "draw walk tiles" on the PathfindingManager and try destroying a pot to observe. It will be marked red (unwalkable) until destroyed, then the tile is marked walkable and enemies can properly path through that tile.

Currently the component is only applied to the Pot spawned object, but will need to be applied to any destroyable spawned object (such as rocks when that is added).