Destroyed Tile Autosave
When tiles are destroyed, they will be autosaved when the room is cleared with other save data. When reloading the save, you can see the tiles correctly remain destroyed and are walkable.

Additionally, I realized since we are using a bitmap setup for pathfinding, we only need one grid. No need for the three separate ones, so I also made that change in this PR.