Base Implementation
The base of the audio system added. The workflow is like this:

- Create a "Sound" scriptable object
- Fill in related settings for the sound
- Instead of setting audio clips to play for attacks etc use Sounds instead
- Play the sound by calling AudioManager.instance.Play(Sound)
- All play logic is handled by the AudioManager/Sound object

Can play oneshots and basic music. The code is messy, uncommented, and incomplete in places but this should provide a solid base moving forward. Coming is sound containers and a music system, as well as cleaning/commenting the code