Sequential and RandomRandom containers working
As of this push the basic sounds, sequential containers and RandomRandom containers are working. Code is still uncommented and disorganized. Next I will implement RandomSequential and RandomOneshot. After that implementation I will add random pitching and volume functionality to the BasicSound, then clean everything up and create a pull request to main