Code clean-up
Added summaries and comments to lines lacking those details, as per code Pull Request review. Added a way to play an AudioClip directly on an Actor's AudioSource, however this is mostly for quality of life changes. Added a warning warning the user to not use the method in this way