projectile trail visuals
These are the last set of visuals I'm going to do for now (or probably at all). I am debating getting rid of the chord visual system in favor of a straight forward text popup (like the damage numbers have) so we don't stretch into areas we know we can't safely stretch into.