Enemy Sounds Update
Big update where all currently available (as of this push) enemy sounds, health sounds, status sounds, and pickup/environmental sounds are implemented. I made a bunch of changes to the audio manager, and created classes to call sounds as part of a state action, and when something is destroyed. Enjoy!